requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
google-cloud-storage
apify-client
openpyxl
requests
python-dotenv
